tests/wpt/web-platform-tests/app-history/navigate/state/navigate-state-after-history-pushState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  appHistory.navigate("#", { replace : true, state : { data : "value" } });
  assert_equals(appHistory.current.getState().data, "value");
  history.pushState(1, "", "#push");
  assert_equals(appHistory.current.getState(), undefined);
}, "appHistoryEntry.getState() after history.pushState()");
</script>
